fix(realtime): prevent sending expired tokens
